refactor: option指定を整理

+++ cmd/pack.go
@@ -2,6 +2,7 @@ package cmd
 
 import (
 	"fmt"
+	"github.com/elecdeer/arduino-esp32-SPIFFS-tool/cmd/options"
 	"github.com/elecdeer/arduino-esp32-SPIFFS-tool/operation/common"
 	"github.com/elecdeer/arduino-esp32-SPIFFS-tool/operation/pack"
 	"github.com/spf13/cobra"
@@ -13,48 +14,53 @@ var packCmd = &cobra.Command{
 	Use:   "pack",
 	Short: "Create a SPIFFS image from files",
 	Args:  cobra.ExactArgs(2),
-	Run:   RunPack,
+	RunE:  RunPack,
 }
 
-func RunPack(cmd *cobra.Command, args []string) {
+func RunPack(cmd *cobra.Command, args []string) error {
 	common.ApplyLogStyle()
-	param, err := ConstructPackParams(args)
+	distPath := args[0]
+	sourceDir := args[1]
+
+	if !common.IsDirExists(sourceDir) {
+		return fmt.Errorf("Source directory not found: %s\n", sourceDir)
+	}
+
+	partition, err := options.ResolvePartitionOptions()
 	if err != nil {
-		cmd.PrintErrf("failed to construct pack params\n%s\n", err)
-		return
+		return fmt.Errorf("Error resolving partition scheme: %s\n", err)
 	}
 
 	tempDir, err := common.MakeTempDir()
 	if err != nil {
-		cmd.PrintErrf("failed to make temp dir\n%s\n", err)
-		return
+		return fmt.Errorf("failed to make temp dir\n%s\n", err)
 	}
 	defer os.RemoveAll(tempDir)
 	cmd.Printf("use temp dir: %s\n", tempDir)
 
 	cmd.Printf("copy files to temp dir\n")
-	err = pack.MakeTargetDir(param.sourceDir, tempDir, param.ignoreDotfile, param.ignorePattern)
+	err = pack.MakeTargetDir(sourceDir, tempDir, packOptions.ignoreDotfile, packOptions.ignorePattern)
 	if err != nil {
-		cmd.PrintErrf("failed to make target dir\n%s\n", err)
-		return
+		return fmt.Errorf("failed to make target dir\n%s\n", err)
 	}
 
 	pack.PrintDirFiles(tempDir)
 
-	err = pack.MakeSpiffsImage(param.mkspiffsPath, param.distPath, param.sourceDir, pack.MakeSpiffsImageOptions{
-		PageSize:  param.pageSize,
-		BlockSize: param.blockSize,
-		Partition: param.partition,
+	err = pack.MakeSpiffsImage(packOptions.packToolPath, distPath, sourceDir, pack.MakeSpiffsImageOptions{
+		PageSize:  packOptions.pageSize,
+		BlockSize: packOptions.blockSize,
+		Partition: &partition,
 	})
 	if err != nil {
-		cmd.PrintErrf("failed to make spiffs image\n%s\n", err)
-		return
+		return fmt.Errorf("failed to make spiffs image\n%s\n", err)
 	}
-	cmd.Printf("SPIFFS image created: %s\n", param.distPath)
+	cmd.Printf("SPIFFS image created: %s\n", distPath)
+
+	return nil
 }
 
 type PackParam struct {
-	partition     *pack.PartitionSchema
+	partition     *options.PartitionSchema
 	mkspiffsPath  string
 	distPath      string
 	sourceDir     string
@@ -64,45 +70,41 @@ type PackParam struct {
 	blockSize     int
 }
 
-func ConstructPackParams(args []string) (PackParam, error) {
-	distPath := args[0]
-	sourceDir := args[1]
+//type FileSystem string
+//
+//const (
+//	LittleFS = FileSystem("littlefs")
+//	Spiffs   = FileSystem("spiffs")
+//)
 
-	if !common.IsFileExists(partitionSchemePath) {
-		return PackParam{}, fmt.Errorf("Partition scheme file not found: %s\n", partitionSchemePath)
-	}
-
-	if !common.IsDirExists(sourceDir) {
-		return PackParam{}, fmt.Errorf("Source directory not found: %s\n", sourceDir)
-	}
-
-	partition, err := pack.ReadPartitionSchemeFile(partitionSchemePath)
-	if err != nil {
-		return PackParam{}, fmt.Errorf("Error reading partition scheme file: %s\n", err)
-	}
-
-	return PackParam{
-		distPath:      distPath,
-		sourceDir:     sourceDir,
-		partition:     &partition,
-		mkspiffsPath:  mkspiffsPath,
-		ignoreDotfile: true,
-		ignorePattern: "",
-		pageSize:      256,
-		blockSize:     4096,
-	}, nil
+type ParsedPackOptions struct {
+	packToolPath        string
+	partitionSchemePath string
+	//fileSystem          FileSystem
+	ignoreDotfile bool
+	ignorePattern string
+	blockSize     int
+	pageSize      int
 }
 
 var (
-	partitionSchemePath string
-	mkspiffsPath        string
+	packOptions ParsedPackOptions
 )
 
 func init() {
 	rootCmd.AddCommand(packCmd)
 
-	packCmd.Flags().StringVarP(&partitionSchemePath, "partition-scheme", "p", "", "Partition scheme csv file path")
-	packCmd.MarkFlagRequired("partition-scheme")
+	packCmd.Flags().StringVarP(&packOptions.packToolPath, "tool", "t", "", "Pack tool path")
+	packCmd.MarkFlagRequired("tool")
 
-	packCmd.Flags().StringVarP(&mkspiffsPath, "mkspiffs", "m", "", "mkspiffs path")
+	options.AddPartitionOptions(packCmd)
+	AddPackCommandOptions(packCmd)
+}
+
+func AddPackCommandOptions(cmd *cobra.Command) {
+	cmd.Flags().BoolVar(&packOptions.ignoreDotfile, "dotfile", true, "Ignore dotfile")
+	cmd.Flags().StringVarP(&packOptions.ignorePattern, "ignore", "i", "", "Ignore file regexp pattern")
+
+	cmd.Flags().IntVar(&packOptions.blockSize, "block", 4096, "Block size")
+	cmd.Flags().IntVar(&packOptions.pageSize, "page", 256, "Page size")
 }
